pyproject.toml
[project]
name = "lab-rag"
version = "0.1.0"
description = "Building a RAG pipeline from scratch"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "nltk==3.9.1",
    "tqdm>=4.67.3",
    "sentence-transformers>=5.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
pythonpath = ["."]
